% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_backward_weights.R
\name{fbweights}
\alias{fbweights}
\title{Forward-backward algorithm}
\usage{
fbweights(PHap, CHap, d, epsilon = 1e-08)
}
\arguments{
\item{PHap}{Named list of parental haplotypes which must be of the form
\code{list(m=, f=)}.\code{m} is a (0,1)-matrix of maternally-inherited haplotypes and
\code{f} is a (0,1)-matrix of paternally-inherited haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{CHap}{(0,1)-matrix of offspring haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{d}{Vector of genetic distances corresponding to haplotypes in PHap and CHap}

\item{epsilon}{De novo mutation rate (default is 1e-8)}
}
\value{
Named list of forward-backward weights
}
\description{
Computes forward-backward weights via the forward-backward algorithm
}
\examples{

}
