% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_clusters}
\alias{make_clusters}
\title{Make clusters for parallel processing}
\usage{
make_clusters(cores = 1)
}
\arguments{
\item{cores}{Number of threads to be used for parallel processing.}
}
\value{
A cluster object.
}
\description{
This function takes a number of cores and creates the corresponding
number of clusters.
}
