% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regions.R
\name{make_regions}
\alias{make_regions}
\title{Makes conditioning regions around SNPs}
\usage{
make_regions(map, snps, mb)
}
\arguments{
\item{map}{A data frame containing a genetic map.}

\item{snps}{A vector of rsIDs for the instruments.}

\item{mb}{Distance in megabases from each instrument to start conditioning on chromosome.}
}
\value{
A list of conditioning regions.
}
\description{
This function takes a genetic map, list of instruments (rsIDs) and
distance (in megabases) and returns a list of conditioning regions consisting of:
\enumerate{
\item the instruments included and 2) the lower and upper boundaries in base positions.
}
}
