% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_score.R
\name{prop_score}
\alias{prop_score}
\title{Propensity scores}
\usage{
prop_score(PHap, CHap, Jset, d, epsilon = 1e-08)
}
\arguments{
\item{PHap}{Named list of mother or father haplotypes}

\item{CHap}{Matrix of offspring haplotypes}

\item{Jset}{List of indices for boundary and instrument SNPs}

\item{d}{Vector of genetic distances corresponding to haplotypes in PHap and CHap}

\item{epsilon}{De novo mutation rate (default is 1e-8)}
}
\value{
Matrix of propensity scores
}
\description{
Computes propensity scores for the instruments
}
\examples{

}
