% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_haplotype.R
\name{sample_haplotype}
\alias{sample_haplotype}
\title{Haplotype sampler}
\usage{
sample_haplotype(prob)
}
\arguments{
\item{prob}{Sampling probabilities from \code{prop_score} function}
}
\value{
Vector of offspring haplotypes
}
\description{
Samples a counterfactual offspring haplotype
}
