% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Haplotype sampler}
\usage{
sampler(N, prob)
}
\arguments{
\item{N}{Sample size}

\item{prob}{Sampling probabilities from \code{prop_score} function}
}
\value{
Vector of offspring haplotypes
}
\description{
Samples a counterfactual offspring haplotype
}
\examples{

}
