% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_backward_weights.R
\name{fbweights}
\alias{fbweights}
\title{Forward-backward algorithm}
\usage{
fbweights(PHap, CHap, d, epsilon = 1e-08)
}
\arguments{
\item{PHap}{Named list of mother or father haplotypes}

\item{CHap}{Matrix of offspring haplotypes}

\item{d}{Vector of genetic distances corresponding to haplotypes in PHap and CHap}

\item{epsilon}{De novo mutation rate (default is 1e-8)}
}
\value{
Named list of forward-backward weights
}
\description{
Computes forward-backward weights via the forward-backward algorithm
}
\examples{

}
