% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_score.R
\name{prop_score}
\alias{prop_score}
\title{Propensity scores}
\usage{
prop_score(PHap, CHap, Jset, d, epsilon = 1e-08)
}
\arguments{
\item{PHap}{Named list of parental haplotypes which must be of the form
\code{list(m=, f=)}.\code{m} is a (0,1)-matrix of maternally-inherited haplotypes and
\code{f} is a (0,1)-matrix of paternally-inherited haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{CHap}{(0,1)-matrix of offspring haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{Jset}{List of indices for boundary and instrument SNPs.}

\item{d}{Vector of genetic distances corresponding to haplotypes in PHap and CHap}

\item{epsilon}{De novo mutation rate (default is 1e-8)}
}
\value{
Matrix of propensity scores
}
\description{
Computes propensity scores for the instruments
}
\examples{

}
