% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_score.R
\name{prop_score}
\alias{prop_score}
\title{Propensity scores}
\usage{
prop_score(PHap, CHap, map, region, epsilon = 1e-08)
}
\arguments{
\item{PHap}{Named list of parental haplotypes which must be of the form
\code{list(m=, f=)}.\code{m} is a (0,1)-matrix of maternally-inherited haplotypes and
\code{f} is a (0,1)-matrix of paternally-inherited haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{CHap}{(0,1)-matrix of offspring haplotypes. Each row corresponds
to an offspring and each column is a SNP.}

\item{map}{Map file of genetic distances between each adjacent SNP. Must contain a
variable called dist.}

\item{epsilon}{De novo mutation rate (default is 1e-8)}

\item{mb}{Distance in megabases from each instrument to start conditioning on chromosome.}

\item{rsids}{rsID(s) of the instrument(s).}
}
\value{
Matrix of propensity scores
}
\description{
Computes propensity scores for the instruments
}
