% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_test.R
\name{run_test}
\alias{run_test}
\title{Main randomisation inference test}
\usage{
run_test(
  reps,
  beta = 0,
  dat,
  prob,
  ins,
  sig = 0.05,
  nnodes = 1,
  out = c("pvalues", "ci", "tobs"),
  verbose = T
)
}
\arguments{
\item{reps}{Number of resamples}

\item{beta}{Vector or scalar of null hypotheses (default is 0)}

\item{dat}{List of data}

\item{prob}{Sampling probabilities from \code{prop_score} function}

\item{ins}{Matrix of instruments}

\item{sig}{Significance level (default is 0.05)}

\item{nnodes}{Number of cores to use during parallel computing (default is 1)}

\item{out}{List of objects to return from the function (default is main statistics)}

\item{verbose}{Choose whether to show the progress bar (default is TRUE)}
}
\value{
F-statistic for the instruments in the linear model
}
\description{
For a set of null hypotheses, computes the observed test statistic,
null distribution of test statistics, p-values and confidence intervals
}
\examples{

}
