% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{test_stat}
\alias{test_stat}
\title{Test statistic}
\usage{
test_stat(beta, pheno, ins)
}
\arguments{
\item{beta}{Vector or scalar of null hypotheses to be tested.}

\item{pheno}{List of phenoa which must be of the form \code{list(exp=, out=, cov=)}.
\code{exp} is a vector containing the exposure variable, \code{out} is a vector
containing the outcome variable and \code{cov} is a matrix containing the covariates
to be included in the test statistic.}

\item{ins}{Matrix of instruments.}
}
\value{
F-statistic for the instruments in the linear model
}
\description{
Computes the test statistic as the F-statistic for the instruments
from a linear regression of the adjusted outcome on the instruments and a user-specified
set of covariates.
}
