% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_stat.R
\name{test_stat}
\alias{test_stat}
\title{Test statistic}
\usage{
test_stat(beta, dat, ins)
}
\arguments{
\item{beta}{Vector or scalar of null hypotheses}

\item{dat}{List of data}

\item{ins}{Matrix of instruments}
}
\value{
F-statistic for the instruments in the linear model
}
\description{
Test statistic
}
\examples{

}
